instance handling and component splitting second stab
