--- src/components/searchresult/yanked.html
@@ -1,72 +1,0 @@
-
-<div class="resultInfo">Time</div>
-          <div class="timeTitle" />
-          <div class="resultDate">
-            Rapport from approx <span>{{ this.transformDate(result.doclist.docs[0].ctime) }} </span>
-            years ago, {{ this.deliverTimeBetween(result.doclist.docs[0].ctime) }}
-            <hr class="report-divider" align="left" />
-            <div v-if="result.doclist.docs[0].ff_primaryobject_year_from_i && result.doclist.docs[0].ff_primaryobject_year_from_i">
-              Rapport content from time period between
-              <span :title="this.returnACorDC(result.doclist.docs[0].ff_primaryobject_year_from_i)">&nbsp; 
-              {{ this.deliverTimePeriodStamp(result.doclist.docs[0].ff_primaryobject_year_from_i) }}</span> and 
-              <span :title="this.returnACorDC(result.doclist.docs[0].ff_primaryobject_year_to_i)">&nbsp;
-              {{ this.deliverTimePeriodStamp(result.doclist.docs[0].ff_primaryobject_year_to_i) }}</span>
-            </div>
-          </div>
-          <div class="placeTitle">Place</div>
-          <div class="resultPlace">{{ this.returnAddresses(result.doclist.docs[0].place_name) }}</div>
-         
-         
-         
-         //Map
-           <div class="mapContainer">
-          <result-map
-            :id="result.doclist.docs[0].id"
-            :coordinateSet="result.doclist.docs[0].place_coordinates || 'Unknown'"
-          />
-        </div> 
-
-
-          <!--snippets-->
-          <div class="seeAllSnippetsBottomContainer">
-            <span v-if="this.showingAllSnippets" class="seeAllSnippets" v-on:click="toggleMoreSnippetsVisibility()">
-                <span>See less hits</span>
-            </span>
-        </div>
-
-
-
-
-
-         <!--PDF Search result handling-->
-        <div :v-if="!queryString.includes('&pt=')" class="matches">
-          <span class="numbersFound">{{ result.doclist.numFound }}</span>
-          <span v-if="result.doclist.numFound > 1"> matches </span>
-          <span v-if="result.doclist.numFound === 1"> match </span>
-          <span>found in pdf. Displaying </span>
-          <span v-if="result.doclist.docs.length <= defaultVisibleSnippets" class="numbersFound">{{ this.result.doclist.docs.length }}</span>
-          <span v-if="result.doclist.docs.length > defaultVisibleSnippets" class="numbersFound">{{ defaultVisibleSnippets }}</span>.
-          <span v-if="result.doclist.docs.length > defaultVisibleSnippets" class="seeAllSnippets" v-on:click="toggleMoreSnippetsVisibility()">
-          <span v-if="showingAllSnippets === true">See less hits</span>
-          <span v-if="showingAllSnippets === false">See more hits</span>
-          </span>
-        </div>
-        <div v-if="this.queryString.includes('&pt=') !== true">
-          <div v-for="(snippets, index) in result.doclist.docs"
-            class="snippet"
-            v-bind:key="index"
-            :style="index > defaultVisibleSnippets - 1 && !showingAllSnippets ? 'display:none' : 'display:block'">
-            <div class="chapterTitle">chapter </div>
-            <highlighted-chapter :chapterString="snippets.chapter" :query="queryString" />
-            <div class="pageTitle">page </div>
-            <div class="pageNumber">{{snippets.page == 0 ? 1 : snippets.page - 1}}</div>
-            <ul>
-            <highlighted-content :contentArray="snippets.highLightSnippets" :query="queryString" />
-            </ul>
-            <router-link :to="getRecordLink(snippets.id, true, snippets.loar_id)">
-              <span>⮊</span> Go to hit
-            </router-link>
-        </div>
-      <router-link class="entirePdfLink" :to="this.getRecordLink(this.result.doclist.docs[0].id, false, this.result.doclist.docs[0].loar_id)">
-        See entire pdf
-      </router-link>
